+++ chevah/compat/administration.py
@@ -23,6 +23,7 @@ from __future__ import absolute_import
 from builtins import str
 from builtins import range
 from builtins import object
+from future.utils import native
 from contextlib import contextmanager
 import os
 import codecs
@@ -30,7 +31,6 @@ import random
 import socket
 import subprocess
 import sys
-import traceback
 
 from chevah.compat import (
     LocalFilesystem,
@@ -41,18 +41,20 @@ from chevah.compat import (
 from chevah.compat.winerrors import ERROR_NONE_MAPPED
 
 
-def execute(
-    command, input_text=None, output=None, ignore_errors=True, verbose=False,
-        ):
+def execute(command, input_text=None, output=None, ignore_errors=True):
     """
     Execute a command having stdout redirected and using 'input_text' as
     input.
     """
+    verbose = False
+
     if verbose:
         print('Calling: %s' % command)
 
     if output is None:
         output = subprocess.PIPE
+
+    command = [native(part) for part in command]
 
     process = subprocess.Popen(
         command, stdin=subprocess.PIPE, stdout=output)
@@ -62,8 +64,6 @@ def execute(
     if exit_code != 0:
         if verbose:
             print(u'Failed to execute %s\n%s' % (command, stderrdata))
-
-        print(''.join(traceback.format_stack(limit=5)))
 
         if not ignore_errors:
             sys.exit(exit_code)
@@ -168,6 +168,14 @@ class OSAdministrationUnix(object):
             '-n', group_name,
             ])
 
+    def _addGroup_openbsd(self, group):
+        group_name = group.name.encode('utf-8')
+        execute([
+            'sudo', 'groupadd',
+            '-g', str(group.gid),
+            group_name,
+            ])
+
     def addUsersToGroup(self, group, users=None):
         """
         Add the users to the specified group.
@@ -225,6 +233,12 @@ class OSAdministrationUnix(object):
         execute([
             'sudo', 'pw', 'groupmod', group_name,
             '-M', members_list.encode('utf-8')])
+
+    def _addUsersToGroup_openbsd(self, group, users):
+        group_name = group.name.encode('utf-8')
+        for user in users:
+            execute([
+                'sudo', 'usermod', '-G', group_name, user.encode('utf-8')])
 
     def addUser(self, user):
         """
@@ -384,6 +398,30 @@ class OSAdministrationUnix(object):
             execute([
                 'sudo', 'chgrp', user.home_group.encode('utf-8'), home_path])
 
+    def _addUser_openbsd(self, user):
+        home_path = user.posix_home_path.encode('utf-8')
+        command = [
+            'sudo', 'useradd',
+            '-u', str(user.uid).encode('utf-8'),
+            '-d', home_path,
+            '-s', user.shell.encode('utf-8'),
+            ]
+
+        if user.posix_home_path != '/tmp':
+            command.append('-m'),
+
+        # Only add gid if required.
+        if user.uid != user.gid:
+            command.extend(['-g', str(user.gid)])
+
+        command.append(user.name.encode('utf-8'))
+
+        execute(command)
+
+        if user.home_group:
+            execute([
+                'sudo', 'chgrp', user.home_group.encode('utf-8'), home_path])
+
     def setUserPassword(self, user):
         """
         Set a password for the user. The password is an attribute of the
@@ -498,6 +536,19 @@ class OSAdministrationUnix(object):
             input_text=user.password.encode('utf-8'),
             )
 
+    def _setUserPassword_openbsd(self, user):
+        code, out = execute(
+            command=['encrypt'],
+            input_text=user.password.encode('utf-8'),
+            )
+
+        execute(
+            command=[
+                'sudo',
+                'usermod', '-p', out.strip(), user.name.encode('utf-8'),
+                ],
+            )
+
     def deleteUser(self, user):
         """
         Delete user from the local operating system.
@@ -528,6 +579,9 @@ class OSAdministrationUnix(object):
         self._deleteHomeFolder_unix(user)
 
     def _deleteHomeFolder_freebsd(self, user):
+        self._deleteHomeFolder_unix(user)
+
+    def _deleteHomeFolder_openbsd(self, user):
         self._deleteHomeFolder_unix(user)
 
     def _deleteHomeFolder_unix(self, user):
@@ -572,6 +626,10 @@ class OSAdministrationUnix(object):
         execute(['sudo', 'pw', 'userdel', user.name.encode('utf-8')])
         self.deleteHomeFolder(user)
 
+    def _deleteUser_openbsd(self, user):
+        execute(['sudo', 'userdel', user.name.encode('utf-8')])
+        self.deleteHomeFolder(user)
+
     def deleteGroup(self, group):
         """
         Delete group from the local operating system.
@@ -603,6 +661,9 @@ class OSAdministrationUnix(object):
 
     def _deleteGroup_freebsd(self, group):
         execute(['sudo', 'pw', 'group', 'del', group.name.encode('utf-8')])
+
+    def _deleteGroup_openbsd(self, group):
+        execute(['sudo', 'groupdel', group.name.encode('utf-8')])
 
     def _appendUnixEntry(self, segments, new_line):
         """