+++ release-notes.rst
@@ -2,6 +2,24 @@ Release notes for chevah.compat
 ===============================
 
 
+0.43.3 - 08/05/2017
+-------------------
+
+* Initialize the test case with a non-Unicode drop user name.
+
+
+0.43.2 - 05/05/2017
+-------------------
+
+* Fix OpenBSD/FreeBSD password authentication.
+
+
+0.43.1 - 04/05/2017
+-------------------
+
+* Fix bad shadow change in previous release.
+
+
 0.43.0 - 04/05/2017
 -------------------